requirements.txt
click==7.1.1
Flask==1.1.2
Flask-PyMongo==2.3.0
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
py-bcrypt==0.4
pymongo==3.10.1
Werkzeug==1.0.1
